refactor non-settings models namespace
