+++ app/models/spid-rails/application_record.rb
@@ -1,0 +1,9 @@
+module Spid
+  module Rails
+
+    class ApplicationRecord < ActiveRecord::Base
+      self.abstract_class = true
+    end
+
+  end
+end
